+++ postmarker/django/__init__.py
@@ -1,0 +1,3 @@
+# coding: utf-8
+from .backend import EmailBackend, PostmarkEmailMessage, PostmarkEmailMixin, PostmarkEmailMultiAlternatives  # noqa
+from .signals import post_send, pre_send  # noqa