+++ tests/models/test_triggers.py
@@ -1,15 +1,18 @@
 # coding: utf-8
+import pytest
+
 from postmarker.models.base import ModelManager
-from postmarker.models.triggers import InboundRule
+from postmarker.models.triggers import InboundRule, Tag
 
 
 CASSETTE_NAME = 'triggers'
 
 
-def test_triggers_manager(postmark):
-    inboundrules = postmark.triggers.inboundrules
-    assert isinstance(inboundrules, ModelManager)
-    assert inboundrules.client is postmark
+@pytest.mark.parametrize('name', ('inboundrules', 'tags'))
+def test_triggers_manager(postmark, name):
+    manager = getattr(postmark.triggers, name)
+    assert isinstance(manager, ModelManager)
+    assert manager.client is postmark
 
 
 class TestInboundRulesTriggers:
@@ -27,3 +30,29 @@ class TestInboundRulesTriggers:
     def test_delete(self, postmark):
         rule = postmark.triggers.inboundrules.all()[0]
         assert rule.delete() == 'Rule someone@example.com removed.'
+
+
+class TestTagsTriggers:
+
+    def test_create(self, postmark):
+        tag = postmark.triggers.tags.create('welcome')
+        assert isinstance(tag, Tag)
+        assert str(tag) == 'welcome'
+
+    def test_all(self, postmark):
+        tags = postmark.triggers.tags.all()
+        assert len(tags) == 1
+        assert isinstance(tags[0], Tag)
+
+    def test_get(self, postmark):
+        tag = postmark.triggers.tags.all()[0]
+        assert isinstance(tag.get(), Tag)
+
+    def test_edit(self, postmark):
+        tag = postmark.triggers.tags.create('welcome2')
+        tag.edit(MatchName='blabla')
+        assert tag.MatchName == 'blabla'
+
+    def test_delete(self, postmark):
+        tag = postmark.triggers.tags.all()[0]
+        assert tag.delete() == 'Tag 1616731 removed.'