+++ postmarker/__init__.py
@@ -1,4 +1,4 @@
 # coding: utf-8
 
 
-__version__ = '0.8.0dev0'
+__version__ = '0.8.0'