+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 -----------
 
 - Added an ability to download more than 500 items. `#70`_
+- Added `pre_send` and `post_send` Django signals. `#83`_
 
 0.7.2 - 11.03.2017
 ------------------
@@ -115,6 +116,7 @@ Changelog
 .. _#98: https://github.com/Stranger6667/postmarker/issues/98
 .. _#92: https://github.com/Stranger6667/postmarker/issues/92
 .. _#87: https://github.com/Stranger6667/postmarker/issues/87
+.. _#83: https://github.com/Stranger6667/postmarker/issues/83
 .. _#78: https://github.com/Stranger6667/postmarker/issues/78
 .. _#76: https://github.com/Stranger6667/postmarker/issues/76
 .. _#73: https://github.com/Stranger6667/postmarker/issues/73