+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - Added `pre_send` and `post_send` Django signals. `#83`_
 - Inbound rules triggers API. `#75`_
 - Changed logs output stream to default ``sys.stderr``. `#102`_
+- Tags triggers API. `#74`_
 
 0.7.2 - 11.03.2017
 ------------------
@@ -123,6 +124,7 @@ Changelog
 .. _#78: https://github.com/Stranger6667/postmarker/issues/78
 .. _#76: https://github.com/Stranger6667/postmarker/issues/76
 .. _#75: https://github.com/Stranger6667/postmarker/issues/75
+.. _#74: https://github.com/Stranger6667/postmarker/issues/74
 .. _#73: https://github.com/Stranger6667/postmarker/issues/73
 .. _#72: https://github.com/Stranger6667/postmarker/issues/72
 .. _#71: https://github.com/Stranger6667/postmarker/issues/71